Fixed rotation bug and somewhat figured out player following

+++ gyakorlatok/Character_Movement/include/Environment/player.h
@@ -2,6 +2,7 @@
 #define PLAYER_H_
 
 #include "../Utils/utils.h"
+#include "Environment/load_texture.h"
 
 #include <obj/load.h>
 #include <obj/draw.h>
@@ -41,7 +42,7 @@ typedef struct Player
 void init_player(Player* player);
 
 /* Moves the player around */
-void move(Player* player, double speed_FPS);
+void move(Player* player, float speed_FPS);
 
 /* Moves the player around according to the given values */
 void increase_position(Player* player, float dx, float dy, float dz);
@@ -54,5 +55,6 @@ void get_speed(Player* player);
 
 /* Loads in a model for use */
 void load_player_model(Player* player);
+
 
 #endif /*PLAYER_H_*/
